pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autonomous-metal-intelligance"
version = "0.1.0"
description = "Core ML utilities for autonomous metal intelligance"
readme = "README.md"
requires-python = ">=3.11,<3.12"
license = { text = "MIT" }

authors = [
    { name = "Tanul Kumar Srivastava" }
]

dependencies = [
    "numpy==1.26.4",
    "pandas==2.2.2",
    "plotly==5.24.1",
    "scikit-learn==1.4.2",
    "tensorflow==2.15.1",
    "openpyxl==3.1.2",
    "python-dotenv==1.0.1",
    "tqdm==4.67.3",
    "kaggle==1.5.16",
    "shap==0.44.1",
    "pandas-toon==0.1.0",
    "faiss-cpu==1.8.0",
    "sentence-transformers==2.7.0",
    "langchain>=0.2.14,<0.3.0",
    "langchain-community>=0.2.14,<0.3.0",
    "langgraph>=0.2.20,<0.3.0",
    "langchain-groq==0.1.9",
    "nbformat==5.10.4",
    "kaleido==0.2.1"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["core*"]
